Refactor callback URL handler
Less code in one place and less code overall.
